moves flattenOrgRoots to resource from structure.go
