+++ aws/structure.go
@@ -37,7 +37,6 @@ import (
 	"github.com/aws/aws-sdk-go/service/macie"
 	"github.com/aws/aws-sdk-go/service/mq"
 	"github.com/aws/aws-sdk-go/service/neptune"
-	"github.com/aws/aws-sdk-go/service/organizations"
 	"github.com/aws/aws-sdk-go/service/rds"
 	"github.com/aws/aws-sdk-go/service/redshift"
 	"github.com/aws/aws-sdk-go/service/route53"
@@ -5441,34 +5440,4 @@ func flattenRoute53ResolverRuleTargetIps(targetAddresses []*route53resolver.Targ
 	}
 
 	return vTargetIps
-}
-
-func flattenOrgRoots(roots []*organizations.Root) []map[string]interface{} {
-	if len(roots) == 0 {
-		return nil
-	}
-	var result []map[string]interface{}
-	for _, r := range roots {
-		result = append(result, map[string]interface{}{
-			"id":           aws.StringValue(r.Id),
-			"name":         aws.StringValue(r.Name),
-			"arn":          aws.StringValue(r.Arn),
-			"policy_types": flattenOrgRootPolicyTypeSummaries(r.PolicyTypes),
-		})
-	}
-	return result
-}
-
-func flattenOrgRootPolicyTypeSummaries(summaries []*organizations.PolicyTypeSummary) []map[string]interface{} {
-	if len(summaries) == 0 {
-		return nil
-	}
-	var result []map[string]interface{}
-	for _, s := range summaries {
-		result = append(result, map[string]interface{}{
-			"status": aws.StringValue(s.Status),
-			"type":   aws.StringValue(s.Type),
-		})
-	}
-	return result
 }
